:memo: Update API models
