--- AUTOMINDER_APP_API/.vscode/settings.json
@@ -1,2 +1,0 @@
-{
-}
