+++ autominder-api/.env
@@ -9,3 +9,4 @@ DEBUG=app:*
 TOKEN_SECRET=483ff926b9532678d203ef54496cc2df7d98e612a8bcf4646f4c0e38ee63f128
 
 TOKEN_EXP=30d
+
