+++ autominder-api/app.js
@@ -12,8 +12,6 @@ var app = express();
 
 mongoose.connect();
 
-// console.log("app.js: Localhost connection has been completed!");
-
 app.use(cors());
 app.use(logger('dev'));
 app.use(express.json());
