+++ autominder-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "autominder-app-api",
+  "name": "buho-app-alert-api",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -11,10 +11,10 @@
     "cors": "^2.8.5",
     "debug": "~2.6.9",
     "dotenv": "^16.0.3",
-    "express": "^4.18.2",
+    "express": "~4.16.1",
     "express-validator": "^6.14.2",
-    "jsonwebtoken": "^9.0.0",
-    "mongoose": "^6.11.0",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^6.7.2",
     "morgan": "~1.9.1"
   },
   "devDependencies": {
